Move pairing dialog: - Trigger button moved to main input settings - Source code moved to `gui/input`
